Simplify model imports
