+++ pypsdm/models/__init__.py
@@ -1,0 +1,4 @@
+from .gwr import GridWithResults
+from .primary_data import PrimaryData
+
+__all__ = ["GridWithResults", "PrimaryData"]
