+++ pypsdm/models/input/connector/__init__.py
@@ -1,0 +1,5 @@
+from .lines import Lines
+from .switches import Switches
+from .transformer import Transformers2W
+
+__all__ = ["Lines", "Switches", "Transformers2W"]
