+++ pypsdm/models/input/connector/lines.py
@@ -5,7 +5,7 @@ from pandas import Series
 
 from pypsdm.models.enums import RawGridElementsEnum
 from pypsdm.models.input.connector.connector import Connector
-from pypsdm.models.input.container.mixins import HasTypeMixin
+from pypsdm.models.input.mixins import HasTypeMixin
 
 
 @dataclass(frozen=True)
