+++ pypsdm/models/input/container/__init__.py
@@ -1,0 +1,5 @@
+from .grid import GridContainer
+from .participants import SystemParticipantsContainer
+from .raw_grid import RawGridContainer
+
+__all__ = ["GridContainer", "SystemParticipantsContainer", "RawGridContainer"]
