+++ pypsdm/models/result/container/__init__.py
@@ -1,0 +1,4 @@
+from .grid import GridContainer, GridResultContainer
+from .participants import ParticipantsResultContainer
+
+__all__ = ["GridContainer", "GridResultContainer", "ParticipantsResultContainer"]
