+++ pypsdm/models/result/grid/__init__.py
@@ -1,0 +1,15 @@
+from .connector import ConnectorResult, ConnectorsResult
+from .enhanced_node import EnhancedNodeResult, EnhancedNodesResult
+from .node import NodeResult, NodesResult
+from .transformer import Transformer2WResult, Transformers2WResult
+
+__all__ = [
+    "ConnectorResult",
+    "ConnectorsResult",
+    "EnhancedNodeResult",
+    "EnhancedNodesResult",
+    "NodeResult",
+    "NodesResult",
+    "Transformer2WResult",
+    "Transformers2WResult",
+]
